+++ src/Domain/Manny.php
@@ -26,6 +26,7 @@ final class Manny
     public const TASK_REFILLING_DEUTERIUM_TANK = 'refilling_deuterium_tank';
     public const TASK_TURNING_ON_SCUT_RELAY = 'turning_on_scut_relay';
     public const TASK_IMPROVING_PROBE = 'improving_probe';
+    public const TASK_ASSEMBLING_PROBE = 'assembling_probe';
     public const CARGO_CAPACITY = 0.05;
     public const CONTAINER_SPACE = 0.05;